Refactor config map into sub pkg
