+++ pkg/util/config/configuration_map.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package config
 
 import (
 	"fmt"
